fix: address member dashboard review findings
