+++ apps/web/src/components/dashboard/member-dashboard-view/index.tsx
@@ -1,20 +1,14 @@
 import { getActiveSubscription } from '@interdomestik/domain-membership-billing/subscription';
 import {
   Bell,
-  BriefcaseBusiness,
   Car,
   ChevronRight,
   FileText,
-  Files,
   Handshake,
-  Home,
-  LifeBuoy,
   MapPinned,
-  MoreHorizontal,
   Plane,
   ShieldAlert,
   UserRound,
-  type LucideIcon,
 } from 'lucide-react';
 import { getTranslations } from 'next-intl/server';
 import { redirect } from 'next/navigation';
@@ -25,6 +19,7 @@ import { DocumentVaultSummary } from './document-vault-summary';
 import { getRoleRedirect } from './helpers';
 import { resolveMemberHomeHero, type MemberHomeHeroModel } from './hero-resolver';
 import { MainServiceCard } from './main-service-card';
+import { MobileBottomNav } from './mobile-bottom-nav';
 import { NextStepCard, type NextStepModel } from './next-step-card';
 import { PrimaryActionPanel } from './primary-action-panel';
 import { TrustStrip } from './trust-strip';
@@ -117,7 +112,7 @@ export async function MemberDashboardView({ data, locale }: Readonly<MemberDashb
         <MemberTopBar isActive={hasAssistanceAccess} t={t} />
 
         <div className="space-y-1.5 md:space-y-4" data-testid="member-dashboard-priority-region">
-          <MemberHero hero={hero} isActive={isActive} t={t} />
+          <MemberHero hero={hero} isActive={hasAssistanceAccess} t={t} />
 
           <div className="hidden md:block">
             <NextStepCard nextStep={nextStep} t={t} />
@@ -165,10 +160,8 @@ export async function MemberDashboardView({ data, locale }: Readonly<MemberDashb
                   <div key={service.key} data-testid="member-service-ecosystem-card">
                     <MainServiceCard
                       key={service.key}
-                      description={t(`services.cards.${service.key}.body`)}
                       href={service.href}
                       icon={service.icon}
-                      label={t('services.open')}
                       mobileLabel={t(`services.cards.${service.key}.mobile`)}
                       situation={t(`services.cards.${service.key}.situation`)}
                       testId={service.testId}
@@ -385,79 +378,6 @@ function MemberHero({
         </div>
       </div>
     </section>
-  );
-}
-
-function MobileBottomNav({ locale, t }: { locale: string; t: DashboardTranslator }) {
-  const items = [
-    { href: `/${locale}/member`, icon: Home, label: t('bottomNav.home') },
-    { href: `/${locale}/member/claims`, icon: BriefcaseBusiness, label: t('bottomNav.cases') },
-    { href: `/${locale}/member/help`, icon: LifeBuoy, label: t('bottomNav.help') },
-    { href: `/${locale}/member/documents`, icon: Files, label: t('bottomNav.documents') },
-    { href: `/${locale}/member/membership`, icon: MoreHorizontal, label: t('bottomNav.more') },
-  ];
-
-  return (
-    <nav
-      aria-label={t('bottomNav.label')}
-      className="fixed inset-x-0 bottom-0 z-40 border-t border-slate-200 bg-white px-2 pb-[max(env(safe-area-inset-bottom),0.5rem)] pt-1 shadow-[0_-8px_30px_rgba(0,0,0,0.04)] md:hidden"
-      data-testid="mobile-bottom-nav"
-    >
-      <div className="mx-auto grid max-w-md grid-cols-5 gap-1">
-        {items.map((item, index) => (
-          <MobileNavItem
-            key={item.href}
-            href={item.href}
-            icon={item.icon}
-            isActive={index === 0}
-            isPrimary={index === 2}
-            label={item.label}
-          />
-        ))}
-      </div>
-    </nav>
-  );
-}
-
-function MobileNavItem({
-  href,
-  icon: Icon,
-  isActive,
-  isPrimary = false,
-  label,
-}: {
-  href: string;
-  icon: LucideIcon;
-  isActive: boolean;
-  isPrimary?: boolean;
-  label: string;
-}) {
-  if (isPrimary) {
-    return (
-      <a
-        href={href}
-        className="relative flex min-h-12 flex-col items-center justify-center gap-1 rounded-2xl px-1 text-[0.65rem] font-bold leading-none text-emerald-900 transition focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-emerald-900 focus-visible:ring-offset-2 min-[380px]:min-h-14 min-[380px]:px-1"
-      >
-        <span className="absolute -top-6 inline-flex h-12 w-12 items-center justify-center rounded-full bg-emerald-800 text-white shadow-[0_8px_16px_rgba(4,120,87,0.3)] ring-[5px] ring-white transition-transform active:scale-95 min-[380px]:-top-7 min-[380px]:h-14 min-[380px]:w-14">
-          <Icon className="h-6 w-6 min-[380px]:h-7 min-[380px]:w-7" aria-hidden="true" />
-        </span>
-        <span className="mt-8 block max-w-full truncate min-[380px]:mt-9">{label}</span>
-      </a>
-    );
-  }
-
-  return (
-    <a
-      href={href}
-      className={`flex min-h-12 flex-col items-center justify-center gap-1.5 rounded-2xl px-1 text-[0.62rem] font-medium leading-none transition focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-emerald-900 focus-visible:ring-offset-2 min-[380px]:min-h-14 min-[380px]:text-[0.65rem] ${
-        isActive
-          ? 'text-emerald-800 font-bold'
-          : 'text-slate-500 hover:text-emerald-700 hover:bg-slate-50'
-      }`}
-    >
-      <Icon className="h-5 w-5 min-[380px]:h-6 min-[380px]:w-6" aria-hidden="true" />
-      <span className="block max-w-full truncate">{label}</span>
-    </a>
   );
 }
